refactor(llm): extraire les scénarios du chat Territoire en builders (PIL-1547)
Le composant BoutonSyntheseTerritoire est unique et accepte les
scénarios en prop. Les deux variantes (DITP/admin et Coordinateur)
sont des fonctions pures dans scenariosTerritoire.ts. Supprime
BoutonSyntheseTerritoireCoordinateur (duplication du shell de chat).

Aussi: doc askAITerritoireEmails précise la séparation par virgules.
